Demangle CUDA kernel names
